added validation between accounts

+++ app/Services/AdminService.php
@@ -63,22 +63,28 @@ class AdminService
     {
         $user = User::query()->where('phone', $request['phone'])->first();
         if(!is_null($user)) {
-            if(!Auth::attempt($request->only(['phone', 'password']))) {
+            if ($user->hasRole('admin')) {
+                if (!Auth::attempt($request->only(['phone', 'password']))) {
+                    $user = [];
+                    $message = 'phone or password is wrong';
+                    $code = 401;
+                } else {
+                    $user = $this->appendRolesAndPermissions($user);
+                    $user['token'] = $user->createToken("Access Token")->plainTextToken;
+                    $message = 'user logged in successfully';
+                    $code = 200;
+                }
+            } else {
                 $user = [];
-                $message = 'phone or password is wrong';
-                $code = 401;
-            }
-            else {
-                $user = $this->appendRolesAndPermissions($user);
-                $user['token'] = $user->createToken("Access Token")->plainTextToken;
-                $message = 'user logged in successfully';
-                $code = 200;
+                $message = 'not and admin';
+                $code = 403;
             }
         }
         else {
             $message = 'user not found';
             $code = 404;
         }
+
         return ['user' => $user, 'message' => $message, 'code' => $code];
     }
     public function logout(): array
@@ -86,13 +92,19 @@ class AdminService
         $user = Auth::user();
 
         if(!is_null($user)) {
-            $user->tokens()->where('id', $user->currentAccessToken()->id)->delete();
-            $message = 'user logged out successfully';
-            $code = 200;
+            if ($user->hasRole('admin')) {
+                $user->tokens()->where('id', $user->currentAccessToken()->id)->delete();
+                $message = 'user logged out successfully';
+                $code = 200;
+            } else {
+                $user = [];
+                $message = 'not an admin';
+                $code = 403;
+            }
         }
         else {
             $message = 'invalid token';
-            $code = 401;
+            $code = 404;
         }
 
         return ['user' => $user, 'message' => $message, 'code' => $code];
@@ -102,16 +114,21 @@ class AdminService
         $user = Auth::user();
 
         if(!is_null($user)) {
-            if(!Hash::check($request['current_password'], $user['password'])) {
-                $message = 'password is incorrect';
-                $code = 401;
-            }
-            else {
-                $user['password'] = Hash::make($request['password']);
-                $user->save();
-                $user = $this->appendRolesAndPermissions($user);
-                $message = 'password updated successfully';
-                $code = 200;
+            if ($user->hasRole('admin')) {
+                if (!Hash::check($request['current_password'], $user['password'])) {
+                    $message = 'password is incorrect';
+                    $code = 401;
+                } else {
+                    $user['password'] = Hash::make($request['password']);
+                    $user->save();
+                    $user = $this->appendRolesAndPermissions($user);
+                    $message = 'password updated successfully';
+                    $code = 200;
+                }
+            } else {
+                $user = [];
+                $message = 'not an admin';
+                $code = 403;
             }
         }
         else {
